Разделяет логику приложения на два дополнительных файла
