+++ functions.php
@@ -1,0 +1,6 @@
+<?php
+
+function to_format_currency($price) {
+    return number_format($price, 0, '', ' ');
+}
+
